POC performance of a List of inline messages
This commit shows a list of 10 rows where each row contains 1 inline in-app message. Scroll the list up and down to see how responsive the UI is as you scroll.

Note: When you first open the app, the rows of the table are the incorrect height. If you scroll down the list and then back up, the cell heights will get corrected. This issue is probably a sample app List issue and not an issue with the inline in-app message. It's been a while since I've created a UIKit TableView 😅.